requirements.txt
pandas
requests
python-dotenv
mlflow-skinny[databricks]>=3.1.1,<4.0.0
